+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/sources/ValueSource.kt
@@ -1,9 +1,7 @@
 package com.github.ajalt.clikt.sources
 
 import com.github.ajalt.clikt.core.Context
-import com.github.ajalt.clikt.parameters.options.Option
-import com.github.ajalt.clikt.parameters.options.longestName
-import com.github.ajalt.clikt.parameters.options.splitOptionPrefix
+import com.github.ajalt.clikt.parameters.options.*
 
 interface ValueSource {
     data class Invocation(val values: List<String>) {
@@ -36,6 +34,47 @@ interface ValueSource {
             val longestName = option.longestName()
             requireNotNull(longestName) { "Option must have a name" }
             return splitOptionPrefix(longestName).second
+        }
+
+        /**
+         * Create a function that will return string keys for options.
+         *
+         * By default, keys will be equal to the value returned by [name].
+         *
+         * @param prefix A static string prepended to all keys
+         * @param joinSubcommands If null, keys will not include names of subcommands. If given,
+         *   this string be used will join subcommand names to the beginning of keys. For options
+         *   that are in a root command, this has no effect. For option in subcommands, the
+         *   subcommand name will joined. The root command name is never included.
+         * @param uppercase If true, returned keys will be entirely uppercase.
+         * @param replaceDashes `-` characters in option names will be replaced with this character.
+         */
+        fun getKey(
+                prefix: String = "",
+                joinSubcommands: String? = null,
+                uppercase: Boolean = false,
+                replaceDashes: String = "-"
+        ): (Context, Option) -> String = { context, option ->
+            var k = name(option).replace("-", replaceDashes)
+            if (joinSubcommands != null) {
+                k = (context.commandNameWithParents().drop(1) + k).joinToString(joinSubcommands)
+            }
+            k = k.replace("-", replaceDashes)
+            if (uppercase) k = k.toUpperCase()
+            prefix + k
+        }
+
+        /**
+         * Create a function that will return string keys that match the key used for environment variables.
+         */
+        fun envvarKey(): (Context, Option) -> String = { context, option ->
+            val env = when (option) {
+                is OptionWithValues<*, *, *> -> option.envvar
+                is FlagOption<*> -> option.envvar
+                else -> null
+            }
+            println("$env, ${inferEnvvar(option.names, env, context.autoEnvvarPrefix)}")
+            inferEnvvar(option.names, env, context.autoEnvvarPrefix) ?: ""
         }
     }
 }